pyproject.toml
[tool.poetry]
name = "data_science_tutorials"
version = "0.1.0"
description = ""
authors = ["Heiko Kulinna <Heikokulinna@gmx.de>"]

[tool.poetry.dependencies]
python = ">=3.8, <3.9"
pandas = "^1.3.4"
numpy = "^1.21.4"
plotly = "^5.4.0"
mlflow = "==1.14.1" 
azure-storage-blob = "==12.7.1"
azure-identity ="^1.7.0"

sklearn = "^0.0"
pyDOE2 = "^1.3.0"
statsmodels = "^0.13.1"
pyarrow = "^6.0.1"
python-dotenv = "^0.19.2"


[tool.poetry.dev-dependencies]
black = "^21.10b0"
flake8 = "^4.0.1"
isort = "^5.8.0"
pre-commit = "^2.12.1"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
